feat(area): Intent to ship area.front
- Implement area.front option
- Split area's test codes as separate file from line's test.

Ref #1543
